+++ axum-macros/tests/from_ref/pass/skip.rs
@@ -1,0 +1,10 @@
+use axum_macros::FromRef;
+
+#[derive(Clone, FromRef)]
+struct AppState {
+    auth_token: String,
+    #[from_ref(skip)]
+    also_string: String,
+}
+
+fn main() {}